func/requirements.txt
--extra-index-url https://backend:dc222fa6-452f-4253-a2fd-a52e7ba2e7e9@nexus.sigame.com.br/repository/pypi/simple
etria-logger==0.1.8
Flask==2.1.2
cx-Oracle==8.3.0
mnemosine==0.1.1
python-dotenv==0.20.0
